Rename interceptor interface
