+++ src/ApiCore/Transport/Grpc/LoggingInterceptorInterface.php
@@ -33,26 +33,42 @@
 namespace Google\ApiCore\Transport\Grpc;
 
 /**
- * Temporary class to support an interceptor-like interface until gRPC interceptor support is
- * available.
- *
- * @package Google\ApiCore\Transport\Grpc
+ * LoggingInterceptor is used to add logging to gRPC Unary calls.
  */
-interface UnaryInterceptor
+class LoggingInterceptorInterface implements UnaryInterceptorInterface
 {
+    private $unaryCallLogger;
+
+    /**
+     * LoggingInterceptor constructor.
+     *
+     * @param UnaryCallLogger $unaryCallLogger
+     */
+    public function __construct(UnaryCallLogger $unaryCallLogger)
+    {
+        $this->unaryCallLogger = $unaryCallLogger;
+    }
+
     /**
      * @param $method
      * @param $argument
      * @param array $metadata
      * @param array $options
-     * @param $continuation
-     * @return mixed
+     * @param callable $continuation
+     * @return LoggingUnaryCall
      */
     public function interceptUnaryUnary(
         $method,
         $argument,
         array $metadata,
         array $options,
-        $continuation
-    );
+        callable $continuation
+    ) {
+    
+        $this->unaryCallLogger->logRequest($method, $argument, $metadata, $options);
+        return new LoggingUnaryCall(
+            $continuation($method, $argument, $metadata, $options),
+            $this->unaryCallLogger
+        );
+    }
 }
