+++ tests/ApiCore/Tests/Unit/Transport/GrpcTransportTest.php
@@ -35,7 +35,7 @@ namespace Google\ApiCore\Tests\Unit\Transport;
 use Google\ApiCore\Call;
 use Google\ApiCore\Tests\Unit\TestTrait;
 use Google\ApiCore\Testing\MockGrpcTransport;
-use Google\ApiCore\Transport\Grpc\GapicInterceptor;
+use Google\ApiCore\Transport\Grpc\GapicInterceptorInterface;
 use Google\ApiCore\Transport\GrpcTransport;
 use Google\Auth\FetchAuthTokenInterface;
 use Google\Protobuf\Internal\RepeatedField;
@@ -395,7 +395,7 @@ class GrpcTransportTest extends TestCase
                             'credentials' => null,
                         ],
                         null,
-                        [new GapicInterceptor()]
+                        [new GapicInterceptorInterface()]
                     );
                 },
             ],
@@ -409,7 +409,7 @@ class GrpcTransportTest extends TestCase
                             'credentials' => null,
                         ],
                         null,
-                        [new GapicInterceptor()]
+                        [new GapicInterceptorInterface()]
                     );
                 },
             ],
